\name{FpopPSD-package}
\alias{FpopPSD-package}
\alias{FpopPSD}
\docType{package}
\title{
  Functional Pruning Optimal Partitioning with Penalty Segment Dependent
}
\details{
  \link[FpopPSD]{fpopPSD}, 
  \link[FpopPSD]{opPSD}
}
\description{
  FpopPSD is an extension of FPOP, an exact data segmentation algorithm based on functional pruning. This extension implements a penalty that depends on the size of the segments.
}

\author{
Arnaud Liehrmann <arnaud.lieh@gmail.com>
}
