% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFPOP.R
\name{PsiFPOP}
\alias{PsiFPOP}
\title{PsiFPOP}
\usage{
PsiFPOP(
  y,
  beta,
  alpha,
  sampling_method = "rand_one",
  sampling_method_parameter = 1,
  wt = -1
)
}
\arguments{
\item{y}{A vector of observations.}

\item{beta}{A constant used in the calculation of the penalty.}

\item{alpha}{A constant used in the calculation of the penalty.
(recommended: gamma+beta*log(length(y)))}

\item{sampling_method}{A sampling method for candidates ("rand_one", "last",
"rand_with_replacement", "rand_without_replacement", "all", "only").}

\item{sampling_method_parameter}{A parameter used by the sampling method.}

\item{wt}{A vector of weights associated to the observations.}
}
\value{
A list with the changepoints and the number of intervals/candidates 
at each step.
}
\description{
PsiFPOP is an exact segmentation algorithm that extends 
functionnal pruning ideas from FPOP to a least squares criterion with a 
multiscale penalty described in Verzelen et al. 2020 (arXiv:2010.11470).
}
