% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFPOP.R
\name{PsiOP}
\alias{PsiOP}
\title{PsiOP}
\usage{
PsiOP(y, beta, alpha, wt = -1)
}
\arguments{
\item{y}{a vector of data ordered according to an attribute}

\item{beta}{a coefficient used in the calculation of the penalty (recommended: 2.5)}

\item{alpha}{a constant used in the calculation of the penalty (recommended: 10+2.5*log(length(y)))}

\item{wt}{a vector of weight linked to the data}
}
\value{
a list with the changepoints and the number of intervals/candidates at each step
}
\description{
PsiOP is an exact segmentation algorithm.
}
